cv-agent/requirements.txt
flask>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
reportlab>=4.2.0
openai>=1.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
